Move getGlobalLock to utils
To fix circular dependency
